% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hfbio_list.R
\name{download_hfbio_model}
\alias{download_hfbio_model}
\title{Download a Model File from Hugging Face Hub}
\usage{
download_hfbio_model(model_id, filename = "config.json")
}
\arguments{
\item{model_id}{Character string specifying the model repository id (e.g., "facebook/esm2_t33_650M_UR50D").}

\item{filename}{Character string specifying the file to download (default is "config.json").}
}
\value{
Path to the downloaded file.
}
\description{
This function downloads a specific file from a Hugging Face model repository
using the hfhub package. It is mainly used to retrieve model weights or configuration files.
}
\examples{
path <- download_hfbio_model("facebook/esm2_t33_650M_UR50D", "config.json")
print(path)
}
